fortress enum fixed
fortress stats reset upon starting a new game

+++ SEPR-game-master/SEPRKroy/core/src/com/mozarellabytes/kroy/Entities/FireTruckType.java
@@ -68,7 +68,7 @@ public enum FireTruckType {
      * @param AP the attack points for this type of truck
      *
      */
-    FireTruckType(int maxReserve, int speed, Color trailColour, String name, float range, float AP, float maxHP) {
+    FireTruckType(int maxReserve, float speed, Color trailColour, String name, float range, float AP, float maxHP) {
         this.maxReserve = maxReserve;
         this.maxHP = maxHP;
         this.speed = speed;
